showing alert dailog while deleting and logingout

+++ lib/features/admin/presentation/pages/manage_assignment_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:grad_ease/core/common/widgets/confirmation_dailog.dart';
 import 'package:grad_ease/core/theme/color_pallete.dart';
 import 'package:grad_ease/features/admin/presentation/bloc/admin_assignment/admin_assignment_bloc.dart';
 import 'package:grad_ease/features/admin/presentation/pages/upsert_assignment_screen.dart';
@@ -141,9 +142,17 @@ class _ManageAssignmentScreenState extends State<ManageAssignmentScreen> {
                               ),
                               child: IconButton(
                                   onPressed: () {
-                                    context.read<AdminAssignmnetBloc>().add(
-                                        DeleteTimeAdminAssignmnet(
-                                            assignment: assignment));
+                                    showConfirmationDialog(
+                                      context: context,
+                                      title: 'Delete',
+                                      content:
+                                          'Are you sure you want to delete?',
+                                      onConfirm: () {
+                                        context.read<AdminAssignmnetBloc>().add(
+                                            DeleteTimeAdminAssignmnet(
+                                                assignment: assignment));
+                                      },
+                                    );
                                   },
                                   icon: const Icon(
                                     Icons.delete,
